requirements.txt
selenium
pyautoit
instaloader
clipboard
